Revert to numpy dtype

+++ dipy/io/tests/test_stateful_tractogram.py
@@ -806,8 +806,8 @@ def test_create_from_sft():
 
 def test_init_dtype_dict_attributes():
     sft = load_tractogram(filepath_dix['gs.trk'], filepath_dix['gs.nii'])
-    dtype_dict = {'positions': float,
-                  'offsets': int,
+    dtype_dict = {'positions': np.float32,
+                  'offsets': np.int64,
                   'dpp': {'color_x': np.single,
                           'color_y': np.single,
                           'color_z': np.single},
